% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resdet.R
\name{crs_grid}
\alias{crs_grid}
\title{Make a grid in a crs projection from a vector data set}
\usage{
crs_grid(x, crs = NULL, ...)
}
\arguments{
\item{x}{SpatVector (we're assuming polygons)}

\item{crs}{a coordinate reference string (projstring, WKT, etc)}

\item{...}{arguments res and dim for underlying function}
}
\value{
a SpatRaster, with a 1 for where the polygon was and NA elsewhere
}
\description{
This grid will mask the polygons so that only pixels of interest can be isolated.
}
\details{
With ... can specify 'dim = \if{html}{\out{<npixels>}}' for a side length of the grid, or
'res = \if{html}{\out{<resolution>}}' a value for the grid. If only dim is provided, or neither then
a heuristic is used to find a nice clean extent and resolution for the grid.
}
\examples{
luxshp <- system.file("ex/lux.shp", package="terra", mustWork = TRUE)
lux <- terra::vect(luxshp)
crs_grid(lux)
crs_grid(lux, crs = "+proj=laea")
crs_grid(lux, crs = "+proj=utm +zone=2", res = 1000)
}
